perf(geometry): remove the CDT's O(T)-per-item scans — 991 -> 651 ms on ISSUE_129
The seam-conform pass cost +56% geometry time on CSG-heavy models. Six attempts to
optimise the conform itself all failed (lazy seam map, x-range pruning, CDT result
caching by clone and by move), and instrumenting finally explained why: build_seam_map
5 ms, conform_plans 15 ms, the open-edge scan 5 ms — 20 ms against 1400 ms of CDT.
The conform was never the cost. It merely triggers more CDT work, and the CDT was
carrying O(T)-per-item bookkeeping that the added second pass exposed.

Fixing the CDT makes the whole pipeline faster, main included:

  ISSUE_129_N1540... geometryMs, interleaved on an idle machine
    main    991   [1038, 1026, 1008, 1032, 1021]
    branch  651   [689, 668, 692, 681, 722]     -34%

That is faster than main WITH the conform running, not merely regression-free.
AC20-FZK-Haus is unchanged (5 ms).

Correctness untouched — every census metric identical to before this commit:
torn void hosts 199, TOTAL unmatched edges 17792, closed-solid defects 41,
non-invariant 133, snap-collapsed 51. mesh-determinism manifest unchanged.

Split to satisfy the module-size ratchet rather than raise a budget:
cdt.rs 1866 -> 1607 (cdt_tests.rs exempt, cdt/predicates.rs), and
conform.rs 465 -> 380 (conform/emit.rs). All new files under 400 lines.

Ledger entry recorded in scripts/perf/README.md per AGENTS.md: the lesson is that
every prior lever targeted the 20 ms rather than the 1400 ms, which is why they all
measured flat.
